refactor(style): package was reorganized for easier navigation
