+++ core/reaction/amendment.py
@@ -16,7 +16,7 @@ from ..model_info import (
     map_reaction_ids_to_participant_ids,
 )
 from .amendment_config import CofactorConfig, ConvergenceConfig, MatchingConfig
-from .matching import map_reactions_to_kegg_with_relaxation
+from .relaxation_workflow import map_reactions_to_kegg_with_relaxation
 from .scoring import (
     SimilarityCalculator,
     TextNormalizer,
