+++ core/reaction/hierarchy_relaxation.py
@@ -940,110 +940,3 @@ def select_relaxations_by_global_improvement(
             to_relax.append(s)
 
     return to_relax
-
-
-def unified_reaction_objective(
-    base_score: float,
-    relaxation_levels: Optional[Mapping[str, Any]],
-    *,
-    lam: float = 0.1,
-    max_relax_level: int = 1,
-) -> float:
-    """
-    Single objective for ranking KEGG reaction matches: similarity minus relaxation penalty.
-
-    ``base_score`` must be the raw similarity (unchanged). All ranking / top-k selection
-    should use only this return value.
-
-    Args:
-        base_score: Raw reaction similarity (e.g. from ``score_model_against_kegg_reaction``).
-        relaxation_levels: Entity id (e.g. model species id) → relaxation level. If None or
-            empty, or ``lam == 0``, returns ``base_score`` unchanged.
-        lam: Penalty weight λ.
-        max_relax_level: Maximum allowed relaxation level used to normalize penalty
-            to ``[0, 1]``.
-    """
-    if lam == 0:
-        return float(base_score)
-    if not relaxation_levels:
-        return float(base_score)
-
-    levels = [int(v) for v in relaxation_levels.values() if v is not None]
-    if not levels:
-        return float(base_score)
-
-    # Max-only penalty (participant-count independent), normalized to [0, 1].
-    max_lvl = max(1, int(max_relax_level))
-    penalty = float(max(levels))
-    normalized_penalty = penalty / float(max_lvl)
-    return float(base_score) - float(lam) * normalized_penalty
-
-
-def unified_reaction_objective_weighted(
-    base_score: float,
-    relaxation_levels: Optional[Mapping[str, Any]],
-    *,
-    weights: Optional[Mapping[str, float]] = None,
-    lam: float = 0.1,
-    max_relax_level: int = 1,
-) -> float:
-    """
-    Weighted penalty: aggregate ``weight(entity) * relaxation_level(entity)`` then apply λ.
-    """
-    if lam == 0:
-        return float(base_score)
-    if not relaxation_levels:
-        return float(base_score)
-
-    if weights is None:
-        weights = {}
-
-    terms: List[float] = []
-    for ent, lvl in relaxation_levels.items():
-        if lvl is None:
-            continue
-        w = float(weights.get(ent, 1.0))
-        terms.append(w * float(int(lvl)))
-
-    if not terms:
-        return float(base_score)
-
-    # Max-only weighted term, normalized to keep penalty bounded.
-    max_lvl = max(1, int(max_relax_level))
-    penalty = float(max(terms))
-    normalized_penalty = penalty / float(max_lvl)
-    return float(base_score) - float(lam) * normalized_penalty
-
-def should_continue_iteration(
-    current_best_score: float,
-    previous_best_score: Optional[float],
-    relaxation_levels: Mapping[str, Any],
-    to_relax: Union[Set[str], Iterable[str]],
-    *,
-    score_tolerance: float = 1e-3,
-) -> bool:
-    """
-    Whether to run another relaxation iteration after scoring.
-
-    Continue while there are entities in ``to_relax`` **or** the aggregate penalized
-    score still moves by at least ``score_tolerance`` vs ``previous_best_score``.
-
-    Stop when ``to_relax`` is empty **and** either ``previous_best_score`` is still
-    the initial sentinel (``None`` or ``-inf``, nothing to relax on first pass) **or**
-    the score change is below tolerance (stable).
-
-    Callers should initialize ``previous_best_score`` to ``float("-inf")`` before
-    the loop and assign it to the current penalized aggregate at the end of each
-    iteration that continues.
-
-    ``relaxation_levels`` is accepted for API symmetry; default rule uses ``to_relax``.
-    """
-    _ = relaxation_levels
-    need_relax = bool(to_relax)
-    if need_relax:
-        return True
-    if previous_best_score is None or previous_best_score == float("-inf"):
-        return False
-    return abs(float(current_best_score) - float(previous_best_score)) >= float(
-        score_tolerance
-    )
